refactor (about - QnA): replace component with mixin
i'm working on a feature that will require me to use some components in some of these, which means i can't just pass the list items in an array as prop.

could've instead used a slot to insert special (or all) QnA content, but that would require repeating their (applicable) style declarations in all QnA components, and the solution for that repetition is a mixin. since there's 0 script in these components and the non-content HTML is tiny, i decided to abandon the template entirely in favor of the mixin