refactor: colocate all page components
took up rich harris's recommendation for component colocation of route-specific components, since the previous approach was almost duplicating the routes folders (and was less readable)

also, renamed the "components" folder to "lib" because i'll make it easier for me to context-switch with my vue project (which i won't switch to this structure because of 1. worse portfolio/open-source palatability 2. that project's orders of magnitude higher complexity (currently, not to mention what's planned), which this structure's rigidity would clash with)