refactor (home page): rephrase design disclaimer
i shifted my focus back to my distillogue project (because i want to reach the proof-of-concept deployment, for jobseeking's sake), so i decided to 1. edit this disclaimer to remove the date (because the epiphany date doesn't really matter when most of the time since was spent elsewise) and 2. restructure the sentence to prioritize (for the reader's attention-span) the reasoning before the plea